# Simplify and optimize the most common methods of 'Headers'