@princerb 2024-10-21
So the tests, I guess then although generators are faster, they are not for good presentation(